+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 ## [Unreleased]
 
+## [0.8.0] - 2023-04-13
 ### Breaking changes
 - If log files in `log > files` field are missing, the gem raise BackupFailed.
 - If log files are optional, use `optional_files` as YAML key instead of `files`.
+
+## [0.7.0] - 2022-12-16
+- MySQL backup support
 
 ## [0.6.0] - 2022-10-27
 - If log files are missing, invoke tar command with only existing files and raise BackupWarning after the backup finished.